+++ engineering/admin/src/app.module.ts
@@ -1,8 +1,8 @@
 import { ClassSerializerInterceptor, Module } from '@nestjs/common';
 import { APP_FILTER, APP_INTERCEPTOR } from '@nestjs/core';
-import { HttpExceptionFilter } from './core/common/filtros/http-exception-filter.filter';
+import { HttpExceptionFilter } from './core/common/filters/http-exception-filter.filter';
 import { EmailBoxModule } from './emailBox/emailBox.module';
-import { TransformInterceptorResponse } from './core/http/transforma-resposta-interceptor';
+import { TransformInterceptorResponse } from './core/infra/http/transforma-resposta-interceptor';
 
 @Module({
   imports: [EmailBoxModule],
